line-bot-pubmed-sending/requirements.txt
functions-framework==3.*
line-bot-sdk
openai==1.3.6
biopython